pyproject.toml
[project]
name = "gmail-gchat-bot"
version = "0.1.0"
description = "FastAPI backend for a Gmail summary Google Chat bot."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0",
  "apscheduler>=3.10.4",
  "edge-tts>=7.2.7",
  "fastapi>=0.115.0",
  "google-api-python-client>=2.170.0",
  "google-auth-oauthlib>=1.2.0",
  "google-cloud-storage>=3.1.0",
  "httpx>=0.27.0",
  "pydantic-ai>=0.0.24",
  "python-dotenv>=1.0.1",
  "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
